Move Place Views into their own package
